Tweak page styles
Inset the interactive elements on the main page to more strongly enforce
the illusion of depth.

The button on the error pages has been made squarer to match the rest of
the styling and has had its lower margin increased to keep the balance
when the slogan is removed from these pages.